DRF/requirements.txt
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cycler==0.11.0
Django==2.2.25
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.1
Faker==12.0.0
fonttools==4.28.5
idna==3.3
importlib-metadata==4.10.0
isort==5.10.1
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
mixer==7.2.0
mypy-extensions==0.4.3
numpy==1.22.1
packaging==21.3
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
scipy==1.7.3
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing-extensions==4.0.1
urllib3==1.26.8
zipp==3.6.0
